Implement a MuJS standard library
This change introduces a MuJS standard library.  This will allow us
to support an increasing amount of ECMAScript functionality without
needing to do it in the runtime or the compiler.

The basic idea is as follows.

In the TypeScript compiler, all uses of ECMAScript standard types and
functions get bound to the TypeScript library headers.  These headers
are in the node_modules/typescript/lib/ directory, and there is one
per ECMAScript "standard" (lib.d.ts for ECMAScript 5, lib.es2015.d.ts
for ECMAScript 6, lib.es5016.d.ts for ECMAScript 7, and so on).  Note
that these libraries are "header only", since the functionality for
these are of course supplied by whatever runtime (V8, Chakra, etc)
actually executes the resulting JavaScript code.

In our case, we have no such luxury.  So instead, we intercept these
bindings and transform them into bindings to a MuJS standard library.
This library is just a MuIL library written in MuJS code.  Only the
set of operations expressable in basic MuIL can be used to implement
these.  In select few cases, we will need runtime intrinsics, but those
will not be specific to MuJS and will go into the basic Mu library and
shared among all MetaMu compilers.

At this point, the MuJS standard library only contains the ECMAScript
Error type, so that we can represent throwing errors in MuJS/MuIL.