Add interface members to prototypes
This change does two things:

First, and foremost, it adds interface members to prototypes.  This
ensures that interface members are available statically on all objects
of types that implement those interfaces.

Second, we permit dynamic loads through LoadLocationExpression when
the `this` object is `dynamic`.  This is a convenient shortcut compared
to creating different LoadDynamicExpressions when the `this` happens to
be of a dynamic type.