Wait for EC2 instance state changes
This change waits for EC2 instance state changes (to running when
creating, and to terminated when deleting), before proceeding.  Note
that we probably want to replace this with custom wait functionality
so that we're entirely in control of the retries, timeouts, logging, etc.
I'll do this soon as part of adding similar waits for security groups.