Add the initial CocoPy compiler scaffolding
This change introduces a new CocoPy compiler.  It will implement a
subset of Python and translate it into Coconut packages and IL.

The project is underneath tools/cocopy/, alongside the cocojs/ compiler
directory, and is organized into two top-level directories: cmd/ contains
the command line tools and lib/ contains the library package code.

At the moment, the compiler processes a single file at a time.  The
Coconut package metadata is loaded identically to how CocoJS will discover
and load package metadata (including searching "upwards" in the filesystem
for package references).

Most of the Coconut package and AST metadata has been projected as Python
classes in the lib/ast and lib/pack modules.  The only missing pieces here
are the token manipulation abstractions which, for Python, will be far less
interesting due to the absence of any strong typing.  (Most likely we'll
end up with a bunch of runtime assertions instead of nice types.)

I've opted to use the pythonparser package because (a) it seems to be well-
designed, (b) it is meant for tools just like this, and (c) the Google
Grumpy compiler uses it (which at least tells me it's a safe-ish bet).

The transformer is still mostly an empty shell of placeholders.