Eliminate conditional initialization
The prior code was trying to be too "clever", skipping initialization
of properties that haven't been given non-None values.  This was to
tidy up serialization and ensure the default JSON/YAML behavior skipped
undefined fields.  Although that basically works, it causes problems
elsewhere in the code, where certain fields would need to be lazily
initialized; this introduces a level of complexity I'd rather not have,
and furthermore, we will most likely need custom serialization anyway.