Add a smattering of Python transformations
This change mostly sketches out placeholders for the various Python
AST nodes and their transformations.  A smattering of them have been
implemented -- basically, those that were "trivial and obvious" upon
the first pass over this -- however, the goal of this checkin really
is to make the broad brush strokes that we will now fill in.

A few minor changes are included, like better assertions, a few
fixups to required/optional parameters, stringification and equality
checking functions, and so on.

Additionally, a bit of the logic for transforming modules is included.
For example, we shove statements into the module initializer, wire
up the entrypoint routine, etc.  For the most part, this is not
functional, however, and will get a bit more robust in the next commit.