Project a smattering of new AWS resource types
This change introduces a bunch of new AWS resource types:

    * aws:iam/Group
    * aws:iam/Policy
    * aws:iam/Role
    * aws:iam/User
    * aws:kms/Key
    * aws:lambda/Function
    * aws:sqs/Queue

None of the associated resource providers are included; this is
an IDL-only projection of the types, their properties, and the
associated API documentation.  Lacking these is blocking me from
making progress on the cross-platform functions project.

It's a little unfortunate to be digging the hole deeper in advance
of pulumi/coconut#133 landing.  However, we are laser-focused on the
pulumi/coconut#134 prototype, so for now I'll just keep going.