Introduce a new cross-cloud package, Mantle
This introduces a new experimental cross-cloud package, Mantle.  This is part
of getting pulumi/coconut#134 up and running.

For now, there isn't much here, except for configuration of the target cloud and
scheduler, plus the ability to create a cross-cloud function.  (OK, it's not truly
cross-cloud yet, since we have only implemented AWS Lambdas.)

Mantle is, of course, a placeholder name.  Since this is a higher level
abstraction than the unopinionated core primitives, the name mantle corresponds
to the Earth's upper layers (and I thought "Crust" would be a rather odd name).